Fix clipping for regular object rendering.
Move camera closer to the screen (40.0) and reduce object vertices scale (16.0) to avoid overflow.
Increase precision of viewport scale multiplication when projecting vertices to the screen.
Clamp directional lighting to 0.
Remove degenerate poly test as it is possible for polygons to be projected to a single scanline.
Clamp edge rasterisation to screen height.